Split SvgGradientServer
